change the licence
changed to apache licence; allowing me the freedom to decide what I want to  do with this codebase. There are some exciting and "difficult" challenges that would be worth investing in tackling that make more sense to explore while offering free use of the software and source code and contributions under an apache licence.